requirements.txt
#transformers
#torch
#torchvision
#datasets
ollama
#langchain
#beautifulsoup4
#faiss-cpu
#speechrecognition
prompt_toolkit
langchain-community
esprima 
tree_sitter 
tree_sitter_languages
docx2txt
unstructured
selenium
whoosh
langchain-chroma
langchain-openai
pymongo